requirements.txt
pandas >= 1.4.2
biopython>=1.79
pytest >= 7.0
coverage >= 6.0
pytest-cov >= 3.0
pytest-mock >= 3.7